Move DB migrations into a submodule
